fix: address the review findings on the subgraph deploy and promote jobs
Three ways the promotion could move or clear the wrong thing:

- a 'Synced:' line the CLI prints in any form this does not recognise was read
  as fully indexed, which moves the 'latest' tag onto a version that is still
  indexing and deletes the one serving traffic. Only '100%' or an explicit
  check mark counts now, and anything else means "not synced yet".
- clearing the flag deleted whatever the variable held. A deployment that
  landed while the promotion was running now keeps its flag instead of being
  dropped silently, never promoted and never reported.
- a tag listing that failed passed for 'no previous version', which reported
  the promotion as done while leaving the old version undeleted. It throws,
  and the flag survives for the next tick.

Nothing bounded a deployment that neither indexed nor failed - a paused
indexer, or a missing GOLDSKY_API_KEY_<ENV> failing the script before it can
report an outcome - so the schedule picked it up every 30 minutes forever. A
deployment now has PROMOTION_DEADLINE_HOURS (24h) to promote, after which the
watch is given up on with an error and the flag is cleared.

Also:

- the production workflow takes the setup action from the ref it is running
  from: its workspace holds the release being deployed, whose '.github' is
  whatever that release shipped - the action may not be in it at all
- the testing workflow falls back to BSNORG_ACTIONS_SECRET when no GH_TOKEN is
  passed, and checks it up front rather than deploying a version it then
  cannot flag, and which nothing would ever promote
- the deployed version reaches 'gh variable set' through the environment, and
  is validated the way the schedule validates it before use
- a variable that does not hold a version annotates and skips that subgraph
  rather than stopping every other network's promotion
- the schedule derives the env key and the variable name from the environment
  and the network instead of carrying them as literal columns
- the promotion installs the subgraph workspace and the root, not all nine
- the '--env' choices move to 'deploy-envs.ts', shared with the Ormi script
- the setup action's 'build' input no longer explains itself with the ./build
  directory, which is not what it produces
