+++ .github/workflows/deploy-testing-subgraph.yaml
@@ -42,9 +42,21 @@ jobs:
       group: subgraph-deploy-testing-${{ matrix.network }}
       cancel-in-progress: false
     steps:
+      # GH_TOKEN is only ever passed by the workflow that calls this one, so a manual run
+      # falls back to the repository secret. Checked up front: a run that cannot flag its
+      # deployment must not deploy one, as nothing would ever promote it.
+      - name: Check the token the deployment is flagged with
+        env:
+          TOKEN: ${{ secrets.GH_TOKEN || secrets.BSNORG_ACTIONS_SECRET }}
+        run: |
+          set -euo pipefail
+          if [ -z "$TOKEN" ]; then
+            echo "::error::Neither the GH_TOKEN passed by the calling workflow nor the BSNORG_ACTIONS_SECRET repository secret is set"
+            exit 1
+          fi
       - uses: actions/checkout@v3
         with:
-          token: ${{ secrets.GH_TOKEN }}
+          token: ${{ secrets.GH_TOKEN || secrets.BSNORG_ACTIONS_SECRET }}
       - uses: ./.github/actions/setup-subgraph
       # Deploys the new version and watches it for a few minutes, long enough to catch a
       # deployment that cannot index at all. The 'latest' tag is left on the version
@@ -57,5 +69,15 @@ jobs:
           GOLDSKY_FATAL_ERROR_WATCH_MINUTES: "5"
       - name: Flag the deployment for promotion
         env:
-          GH_TOKEN: ${{ secrets.GH_TOKEN }}
-        run: gh variable set ${{ matrix.variable }} --repo "$GITHUB_REPOSITORY" --body "${{ steps.deploy.outputs.version }}"
+          GH_TOKEN: ${{ secrets.GH_TOKEN || secrets.BSNORG_ACTIONS_SECRET }}
+          VARIABLE: ${{ matrix.variable }}
+          VERSION: ${{ steps.deploy.outputs.version }}
+        run: |
+          set -euo pipefail
+          # The flag is read back into a job matrix and onto a command line, so never
+          # write one that does not look like the version that was just deployed
+          if [[ ! "$VERSION" =~ ^[A-Za-z0-9.+-]+$ ]]; then
+            echo "::error::'$VERSION' does not look like a deployed version"
+            exit 1
+          fi
+          gh variable set "$VARIABLE" --repo "$GITHUB_REPOSITORY" --body "$VERSION"
